chore(tools): complete first draft of splitting cli from helper functions (#4825)
